👽️ REST Api >> Refactoring express server. Still need a little more cleanup. KNOWN BUG: Route for deleting files/folders not working right now.
