--- server-middleware/asyncMiddleware.js
@@ -1,8 +1,0 @@
-const asyncUtil = (fn) =>
-    function asyncUtilWrap(...args) {
-        const fnReturn = fn(...args)
-        const next = args[args.length - 1]
-        return Promise.resolve(fnReturn).catch(next)
-    }
-
-module.exports = asyncUtil
