+++ server-middleware/middleware/asyncMiddleware.js
@@ -1,0 +1,7 @@
+const asyncUtil = (fn) =>
+    function asyncUtilWrap(req, res, next, ...args) {
+        const fnReturn = fn(req, res, next, ...args)
+        return Promise.resolve(fnReturn).catch(next)
+    }
+
+module.exports = asyncUtil
