+++ server-middleware/routes/index.js
@@ -1,0 +1,13 @@
+// Express
+const express = require('express')
+const router = express.Router()
+
+/* GET users listing. */
+router.all('/', (req, res, next) => {
+    res.json({
+        _status: 'ok',
+        info: 'Endpoint is set up.'
+    })
+})
+
+module.exports = router
